launcher/requirements.txt
altgraph==0.17.4
asttokens==2.4.1
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
customtkinter==5.2.2
darkdetect==0.8.0
debugpy==1.8.1
decorator==5.1.1
eel==0.16.0
ewmh==0.1.6
executing==2.0.1
flask==3.0.3
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
jupyter-client==8.6.2
jupyter-core==5.7.2
markupsafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
portablemc==4.3.0
prompt-toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pygetwindow==0.0.9
pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
pypresence==4.3.0
pyrect==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-xlib==0.33
pyzmq==26.0.3
requests==2.32.3
setuptools==70.0.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.2
wcwidth==0.2.13
websocket==0.2.1
werkzeug==3.0.3
whichcraft==0.6.1
zope-event==5.0
zope-interface==6.4.post2
